api/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
sqlalchemy==2.0.31
nats-py==2.10.0